chore: move other functions to utils
